backend/requirements.txt
# Django backend requirements for React + Django + PostgreSQL + JWT project

Django>=4.2
dj-database-url
whitenoise
gunicorn
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9
python-dotenv>=1.0
django-cors-headers>=4.3